Added tests, some restructuring
