+++ bouter/deprecated.py
@@ -1,0 +1,24 @@
+import warnings
+import functools
+
+from numpy import VisibleDeprecationWarning
+
+
+def deprecated(message=None):
+    if message is None:
+        message = ""
+
+    def decorator(fn):
+        @functools.wraps(fn)
+        def wrapper(*args, **kwargs):
+            warnings.warn(
+                "Function {0} is deprecated. {1}".format(fn.__name__, message),
+                category=VisibleDeprecationWarning,
+                stacklevel=2,
+            )
+
+            return fn(*args, **kwargs)
+
+        return wrapper
+
+    return decorator
