+++ requirements_dev.txt
@@ -6,7 +6,6 @@ flake8==3.7.8
 tox==3.14.0
 coverage==4.5.4
 Sphinx==1.8.5
-twine==1.14.0
 pytest==4.6.5
 pytest-runner==5.1
 pytest-cov
