+++ setup.cfg
@@ -22,4 +22,5 @@ test = pytest
 
 [tool:pytest]
 collect_ignore = ['setup.py']
+addopts = --cov=bouter
