+++ tests/assets/test_dataset/192316_stimulus_log.hdf5
