+++ tox.ini
@@ -1,17 +1,18 @@
 [tox]
-envlist = py35, py36, py37, py38, flake8
+envlist = py36, py37, py38, flake8
 
 [travis]
 python =
     3.8: py38
     3.7: py37
     3.6: py36
-    3.5: py35
 
 [testenv:flake8]
 basepython = python
 deps = flake8
-commands = flake8 bouter tests
+commands =
+    flake8 bouter tests
+    black bouter tests
 
 [testenv]
 setenv =
